requirements.txt
python-sane==2.*
PyYAML==6.*
zebra==0.2
